chore: verify if restaurant exists in restaurants use case

+++ src/use-cases/restaurants/delete-restaurant.ts
@@ -1,4 +1,5 @@
 import { IRestaurantsRepository } from "@/repositories/restaurants-repository";
+import { RestaurantNotFoundError } from "../errors/restaurant-not-found-error";
 
 type DeleteRestaurantUseCaseRequest = {
   id: string
@@ -8,6 +9,13 @@ export class DeleteRestaurantUseCase {
   constructor(private restaurantsRepository: IRestaurantsRepository) {}
 
   async execute({ id }: DeleteRestaurantUseCaseRequest): Promise<void> {
+
+    const restaurantExists = await this.restaurantsRepository.getById(id);
+
+    if (!restaurantExists) {
+      throw new RestaurantNotFoundError();
+    } 
+
     await this.restaurantsRepository.deleteById(id);
   }
 }
